Click here to see question
Write a program which repeatedly reads numbers until the user en- ters “done”. Once “done” is entered, print out the total, count, and average of the numbers. If the user enters anything other than a number, detect their mistake using try and except and print an error message and skip to the next number.