test 2to3 method
Updated to be python 3 compatible while determining clean way to "lint" for py3.

`2to3 -f all -f idioms -f buffer -f set_literal -f ws_comma <file_or_dir>`

prints a diff if anything needs to change.

Note the output still lists "Files needing to be changed" even if there is no diff.
This is confusing, but with no diff it seems to be a bug with the tool.