work around date formatting
In bash the default format for `date` adds an extra space before single digit day numbers.

This becomes problematic when further manipulating that string, as it gets tokenized and the extra space is removed.

This complicates finding dates in logs, manual or otherwise. This snippet was poc to update dotfiles for consistency in date formatting for log creation and searching.